revert opinionated sub-tab implementation
