+++ common/Root.tsx
@@ -42,10 +42,14 @@ export default class Root extends Component<Props, State> {
       <Provider store={store} key={Math.random()}>
         <Router history={history} key={Math.random()}>
           <div>
-            <Route path="/generate" component={GenerateWallet} />
+            <Route exact={true} path="/" component={GenerateWallet} />
             <Route path="/help" component={Help} />
             <Route path="/swap" component={Swap} />
-            <Route path="/account" component={SendTransaction} />
+            <Route path="/account" component={SendTransaction}>
+              <Route path="send" component={SendTransaction} />
+              <Route path="info" component={SendTransaction} />
+            </Route>
+            <Route path="/send-transaction" component={SendTransaction} />
             <Route path="/contracts" component={Contracts} />
             <Route path="/ens" component={ENS} />
             <Route path="/utilities" component={RestoreKeystore} />
@@ -70,7 +74,7 @@ const LegacyRoutes = withRouter(props => {
         history.push('/send-transaction');
         break;
       case '#generate-wallet':
-        history.push('/generate');
+        history.push('/');
         break;
       case '#swap':
         history.push('/swap');
@@ -92,11 +96,9 @@ const LegacyRoutes = withRouter(props => {
 
   return (
     <Switch>
-      <Redirect from="/" to="/generate" exact={true} />
       <Redirect from="/signmsg.html" to="/sign-and-verify-message" />
       <Redirect from="/helpers.html" to="/helpers" />
       <Redirect from="/send-transaction" to="/account/send" />
-      <Redirect from="/sign-and-verify-message" to="/message/sign" />
     </Switch>
   );
 });
